+++ lib/useConditionalEffect.ts
@@ -1,12 +1,14 @@
-import { useEffect } from 'react'
+import { useEffect, useLayoutEffect } from 'react'
 
 import type { Maybe } from '@/utils/types'
 import type { DependencyList, EffectCallback } from 'react'
+
 /**
  * `useEffect` with conditional function
  * @param effect - Imperative function that can return a cleanup function
  * @param deps - If present, effect will only activate if the values in the list change
  * @param condition - The conditional function that effect or not. If return `true` effect, otherwise not.
+ * @param effectHook - Which effect hooks to use
  *
  * @example
  * ```tsx
@@ -19,9 +21,15 @@ import type { DependencyList, EffectCallback } from 'react'
 const useConditionalEffect = (
   effect: EffectCallback,
   deps?: DependencyList,
-  condition?: () => Maybe<boolean>
+  condition?: () => Maybe<boolean>,
+
+  /**
+   * @defaultValue `useEffect`
+   */
+  effectHook?: typeof useEffect | typeof useLayoutEffect
 ): void => {
-  useEffect(() => {
+  const useEffectHook = effectHook ?? useEffect
+  useEffectHook(() => {
     if (typeof condition === 'function' && condition() !== true) return
     return effect()
     // eslint-disable-next-line react-hooks/exhaustive-deps