+++ lib/utils/index.ts
@@ -1,6 +1,38 @@
 /**
  * Check client-side or server-side
+ *
+ * @internal
  */
 const isBrowser = typeof window !== 'undefined'
 
-export { isBrowser }
+/**
+ * Remove deprecated value
+ * @param value - Any array
+ * @returns New array of unique values
+ *
+ * @internal
+ */
+const uniqShallow = <T>(value: T[]): T[] => Array.from(new Set(value))
+
+/**
+ *
+ * @param a - Base array
+ * @param b - Target array
+ * @returns Array of values not in `b` for `a`
+ *
+ * @internal
+ */
+const without = <T>(a: T[], b: T[]): T[] => {
+  const _a = uniqShallow(a)
+  const _b = uniqShallow(b)
+  return _a.reduce((acc, cur) => {
+    const hasSame = _b.some((__b) => __b === cur)
+
+    if (!hasSame) {
+      acc.push(cur)
+    }
+    return acc
+  }, [] as T[])
+}
+
+export { isBrowser, uniqShallow, without }