+++ stories/shared/constants.ts
@@ -1,10 +1,13 @@
-import type { ExtendedArgType } from '@story/shared/types'
+import type { ExtendedArgType, TableArgTypePolyfill } from '@story/shared/types'
+
+const category: TableArgTypePolyfill['category'] = 'args'
+const subcategory0: TableArgTypePolyfill['subcategory'] = '[0]{ options }'
 
 const deps: ExtendedArgType = {
   description:
     'If present, effect will only activate if the values in the list change',
   table: {
-    category: 'args',
+    category,
     subcategory: '[1]',
     type: {
       summary: 'DependencyList | undefined'
@@ -16,7 +19,7 @@ const condition: ExtendedArgType = {
   description:
     'The conditional function that effect or not. If return `true` effect, otherwise not.',
   table: {
-    category: 'args',
+    category,
     subcategory: '[2]',
     type: {
       summary: '(() => Maybe<boolean>) | undefined'
@@ -24,4 +27,4 @@ const condition: ExtendedArgType = {
   }
 }
 
-export { deps, condition }
+export { deps, condition, category, subcategory0 }