+++ tests/utils/server/index_test.ts
@@ -1,0 +1,5 @@
+import { isBrowser } from '@/utils'
+
+describe('isBrowser', () => {
+  it('should return false', () => expect(isBrowser).toBeFalsy())
+})